% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{dxc}
\alias{dxc}
\alias{pcc}
\title{Diagnostic and Procedure Codes}
\usage{
dxc(category, version)

pcc(category, version)
}
\arguments{
\item{category}{a character string of a CCC category.  See Details.}

\item{version}{ICD version 9 or 10}
}
\value{
A character vector.  \code{dxc} return diagnostic codes, \code{pcc} returns
procedure codes.
}
\description{
ICD codes for Complex Chronic Conditions (CCC) categories.
}
\details{
The CCC categories for diagnostic and procedure codes are:
\tabular{lcc}{
category        \tab \code{dxc} \tab \code{pcc} \cr
neuromuscular   \tab      X     \tab      X     \cr
cvd             \tab      X     \tab      X     \cr
respiratory     \tab      X     \tab      X     \cr
renal           \tab      X     \tab      X     \cr
gi              \tab      X     \tab      X     \cr
hemato_immu     \tab      X     \tab      X     \cr
metabolic       \tab      X     \tab      X     \cr
congeni_genetic \tab      X     \tab            \cr
malignancy      \tab      X     \tab      X     \cr
neonatal        \tab      X     \tab            \cr
tech_dep        \tab      X     \tab      X     \cr
transplant      \tab      X     \tab      X     \cr
}

The ICD codes were taken from the SAS macro provided by the reference paper.
}
\examples{
# ICD 9 diagnostic codes for cvd
dxc("cvd", 9)

# ICD 10 diagnostic codes for cvd
dxc("cvd", 10)

}
\references{
Feudtner C, et al. Pediatric complex chronic conditions classification system
version 2: updated for ICD-10 and complex medical technology dependence and
transplantation, BMC Pediatrics, 2014, 14:199, DOI: 10.1186/1471-2431-14-199
}
\author{
Peter DeWitt
}
