% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{comparability}
\alias{comparability}
\title{Multiple Cause of Death (MCOD) file extract}
\format{
A data frame with 65037 rows and 3 variables.
\describe{
  \item{id}{Sequentially assigned patient identifier}
  \item{icd9}{Underlying Cause of Death ICD 9 CM diagnosis code}
  \item{icd10}{Underlying Cause of Death ICD 10 diagnosis code}
}
}
\usage{
comparability
}
\description{
The Center for Disease Control has made publicly available death certificate data.
This data set is a subset of the 2006 MCOD file for decedents aged <= 21 showing just
the underlying cause of death diagnosis code in ICD-9-CM and ICD-10.
}
\details{
See `vignette("pccc-example")` for more details about the MCOD source file.
}
\keyword{datasets}
