% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_codes}
\alias{get_codes}
\title{Get (view) Diagnostic and Procedure Codes}
\usage{
get_codes(version = 9L)
}
\arguments{
\item{version}{and integer value specifying ICD version 9 (default) or 10.}
}
\value{
A list with three elements
\describe{
  \item{version}{ICD version}
  \item{dx}{A list of diagnostic codes.  Each element is for a CCC category
   as noted in the details}
  \item{pc}{A list of diagnostic codes.  Each element is for a CCC category
   as noted in the details}
}
}
\description{
View the ICD, verion 9 or 10, for the Complex Chronic Conditions (CCC)
categories.
}
\details{
The CCC categories for diagnostic and procedure codes are:
 \tabular{lcc}{                                            
 category        \tab \code{dxc} \tab \code{pcc} \cr       
 neuromuscul     \tab      X     \tab      X     \cr       
 cvd             \tab      X     \tab      X     \cr       
 respiratory     \tab      X     \tab      X     \cr       
 renal           \tab      X     \tab      X     \cr       
 gi              \tab      X     \tab      X     \cr       
 hemato_immu     \tab      X     \tab      X     \cr       
 metabolic       \tab      X     \tab      X     \cr       
 congeni_genetic \tab      X     \tab            \cr       
 malignancy      \tab      X     \tab      X     \cr       
 neonatal        \tab      X     \tab            \cr       
 tech_dep        \tab      X     \tab      X     \cr       
 transplant      \tab      X     \tab      X     \cr       
 }

The ICD codes were taken from the SAS macro provided by the reference paper.
}
\references{
Feudtner C, et al. Pediatric complex chronic conditions classification system
version 2: updated for ICD-10 and complex medical technology dependence and
transplantation, BMC Pediatrics, 2014, 14:199, DOI: 10.1186/1471-2431-14-199
}
\author{
Peter DeWitt
}
