% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pccc_icd9_dataset}
\alias{pccc_icd9_dataset}
\title{Randomly Generated ICD 9 Sample Data Set}
\format{A data frame with 1000 rows and 31 variables
\describe{
  \item{id}{Sequentialy assigned patient identifier}
  \item{dx1:dx10}{ICD 9 diagnosis codes; 20\% of values are missing, see
    \url{https://www.cms.gov/Medicare/Coding/ICD9ProviderDiagnosticCodes/codes.html}
    for source of codes.}
  \item{pc1:pc10}{ICD 9 procedure codes; 20\% of values are missing, see
    \url{https://www.cms.gov/Medicare/Coding/ICD9ProviderDiagnosticCodes/codes.html}
    for source of codes.}
  \item{g1:g10}{Random data to simulate other data often present in export of
    patient data with 20\% of values missing.}
}}
\usage{
pccc_icd9_dataset
}
\description{
This dataset was produced from a tool available at
\url{https://github.com/magic-lantern/icd_file_generator}. ICD codes were taken
from CMS.
}
\keyword{datasets}
