% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_helper.R
\name{test_helper}
\alias{test_helper}
\title{Tool to help access internal variables to use in testthat scripts}
\usage{
test_helper(var)
}
\arguments{
\item{var}{bare name of the internal variable to be accessed.}
}
\value{
Object from internal PCCC name space (if it exists)
}
\description{
Tool to help access internal variables to use in testthat scripts
}
