% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccc.R
\name{ccc}
\alias{ccc}
\title{Complex Chronic Conditions (CCC)}
\usage{
ccc(.data, id, dx_cols, pc_cols, icdv = 10)
}
\arguments{
\item{.data}{a \code{data.frame} containing a patient id and all the ICD-9-CM
or ICD-10-CM codes.}

\item{id}{bare name of the column containing the patient id}

\item{dx_cols}{column names with the diagnositic codes, use
\code{\link[dplyr]{vars}} from \code{dplyr} to identify the column(s).}

\item{pc_cols}{the column(s) containing the procedure codes.  Use
\code{\link[dplyr]{vars}} from \code{dplyr} to identify the column(s).}

\item{icdv}{ICD version 9 or 10, defaiults to 10}
}
\description{
Generate CCC and CCC subcategory flags and the number of categories.
}
\details{
The \code{data.frame} passed to the function should be in wide format.
}
\references{
Feudtner C, et al. Pediatric complex chronic conditions classification system
version 2: updated for ICD-10 and complex medical technology dependence and
transplantation, BMC Pediatrics, 2014, 14:199, DOI: 10.1186/1471-2431-14-199
}
\author{
Peter DeWitt
}
