% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccc.R
\name{ccc}
\alias{ccc}
\title{Complex Chronic Conditions (CCC)}
\usage{
ccc(.data, id, dx_cols = NULL, pc_cols = NULL, icdv)
}
\arguments{
\item{.data}{a \code{data.frame} containing a patient id and all the ICD-9-CM
or ICD-10-CM codes.  The \code{data.frame} passed to the function should be
in wide format.}

\item{id}{bare name of the column containing the patient id}

\item{dx_cols}{column names with the diagnostic codes, use
\code{\link[dplyr]{vars}} from \code{dplyr} to identify the column(s).}

\item{pc_cols}{the column(s) containing the procedure codes.  Use
\code{\link[dplyr]{vars}} from \code{dplyr} to identify the column(s).}

\item{icdv}{ICD version 9 or 10}
}
\value{
A \code{data.frame} with a column for the subject id and integer (0
or 1) columns for each each of the categories.
}
\description{
Generate CCC and CCC subcategory flags and the number of categories.
}
\details{
It is recommended that you view the codes defining the CCC via
\code{\link{get_codes}} and make sure that the ICD codes in your data set are
formatted in the same way.  The ICD codes used for CCC are character strings
and *do not* have decimal points.
}
\examples{
eg_data <-
  dplyr::data_frame(id = letters,
                    dx1 = c(NA, NA, sample(get_codes(10)[["hemato_immu", "dx"]], 24)),
                    dx2 = c("A", sample(get_codes(10)[["gi", "dx"]], 25)),
                    dx3 = LETTERS,
                    pc1 = c("B", sample(get_codes(10)[["cvd", "pc"]], 25)),
                    pc2 = LETTERS,
                    other_col = LETTERS)

ccc(eg_data, 
    id,
    dx_cols = dplyr::vars(dplyr::starts_with('dx')),
    pc_cols = dplyr::vars(dplyr::starts_with("pc")),
    icdv = 10)

}
\references{
Feudtner C, et al. Pediatric complex chronic conditions classification system
version 2: updated for ICD-10 and complex medical technology dependence and
transplantation, BMC Pediatrics, 2014, 14:199, DOI: 10.1186/1471-2431-14-199
}
\seealso{
\code{\link{get_codes}} to view the ICD codes used to define the
CCC.
}
\author{
Peter DeWitt
}
