% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccc.R
\name{ccc}
\alias{ccc}
\title{Complex Chronic Conditions (CCC)}
\usage{
ccc(.data, id, dx_cols, pc_cols, icdv = 10)
}
\arguments{
\item{.data}{a \code{data.frame} containing a patient id and all the ICD-9-CM
or ICD-10-CM codes.}

\item{id}{bare name of the column containing the patient id}

\item{dx_cols}{column names with the diagnostic codes, use
\code{\link[dplyr]{vars}} from \code{dplyr} to identify the column(s).}

\item{pc_cols}{the column(s) containing the procedure codes.  Use
\code{\link[dplyr]{vars}} from \code{dplyr} to identify the column(s).}

\item{icdv}{ICD version 9 or 10, defaults to 10}
}
\description{
Generate CCC and CCC subcategory flags and the number of categories.
}
\details{
The \code{data.frame} passed to the function should be in wide format.
}
\examples{
eg_data <-
  dplyr::data_frame(id = letters,
                    dx1 = c(NA, NA, sample(dxc("hemato_immu", 10), 24)),
                    dx2 = sample(dxc("malignancy", 10), 26),
                    dx3 = LETTERS,
                    pc1 = sample(pcc("cvd", 10), 26),
                    pc2 = LETTERS,
                    other_col = LETTERS)

ccc(eg_data, id,
    dx_cols = dplyr::vars(dplyr::starts_with('dx')),
    pc_cols = dplyr::vars(dplyr::starts_with("pc"))) 
}
\references{
Feudtner C, et al. Pediatric complex chronic conditions classification system
version 2: updated for ICD-10 and complex medical technology dependence and
transplantation, BMC Pediatrics, 2014, 14:199, DOI: 10.1186/1471-2431-14-199
}
\author{
Peter DeWitt
}
