% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pccc_icd10_dataset}
\alias{pccc_icd10_dataset}
\title{Randomly Generated ICD 10 Sample Data Set}
\format{
A data frame with 1000 rows and 31 variables.
There is a patient identifier, ten diagnosis codes, ten procedure codes, and
ten "other data" values, specifically:
\describe{
  \item{id}{Sequentially assigned patient identifier}
  \item{dx1}{a ICD 10 diagnosis code}
  \item{dx2}{a ICD 10 diagnosis code}
  \item{dx3}{a ICD 10 diagnosis code}
  \item{dx4}{a ICD 10 diagnosis code}
  \item{dx5}{a ICD 10 diagnosis code}
  \item{dx6}{a ICD 10 diagnosis code}
  \item{dx7}{a ICD 10 diagnosis code}
  \item{dx8}{a ICD 10 diagnosis code}
  \item{dx9}{a ICD 10 diagnosis code}
  \item{dx10}{a ICD 10 diagnosis code}
  \item{pc1}{a ICD 10 procedure codes}
  \item{pc2}{a ICD 10 procedure codes}
  \item{pc3}{a ICD 10 procedure codes}
  \item{pc4}{a ICD 10 procedure codes}
  \item{pc5}{a ICD 10 procedure codes}
  \item{pc6}{a ICD 10 procedure codes}
  \item{pc7}{a ICD 10 procedure codes}
  \item{pc8}{a ICD 10 procedure codes}
  \item{pc9}{a ICD 10 procedure codes}
  \item{pc10}{a ICD 10 procedure codes}
  \item{g1}{Random data to simulate other data often present in export of patient data with 20\% of values missing.}
  \item{g2}{Random data to simulate other data often present in export of patient data with 20\% of values missing.}
  \item{g3}{Random data to simulate other data often present in export of patient data with 20\% of values missing.}
  \item{g4}{Random data to simulate other data often present in export of patient data with 20\% of values missing.}
  \item{g5}{Random data to simulate other data often present in export of patient data with 20\% of values missing.}
  \item{g6}{Random data to simulate other data often present in export of patient data with 20\% of values missing.}
  \item{g7}{Random data to simulate other data often present in export of patient data with 20\% of values missing.}
  \item{g8}{Random data to simulate other data often present in export of patient data with 20\% of values missing.}
  \item{g9}{Random data to simulate other data often present in export of patient data with 20\% of values missing.}
  \item{g10}{Random data to simulate other data often present in export of patient data with 20\% of values missing.}
}
}
\usage{
pccc_icd10_dataset
}
\description{
This dataset was produced from a tool available at
\url{https://github.com/magic-lantern/icd_file_generator}. ICD codes were taken
from CMS.  The code source, for both the diagnosis and produced codes can be
found at \url{https://www.cms.gov/Medicare/Coding/ICD10/2017-ICD-10-CM-and-GEMs.html}
}
\keyword{datasets}
